refactor: migrate bot logic to main.py and update CI/Docker configuration with uv
