+++ .dockerignore
@@ -1,0 +1,2 @@
+.venv
+dist
