+++ Dockerfile
@@ -1,10 +1,17 @@
-FROM python:3.12-slim
+FROM python:3.13-slim-bookworm
 
+# Copy uv from the official image
+COPY --from=ghcr.io/astral-sh/uv:0.7.8 /uv /uvx /bin/
+
+ADD . /app
 WORKDIR /app
 
-COPY . .
-RUN curl -LsSf https://astral.sh/uv/install.sh | sh
-ENV PATH="/root/.cargo/bin:/root/.local/bin:$PATH"
-RUN uv sync
+# Set environment variable for Python to not create .pyc files
+ENV PYTHONDONTWRITEBYTECODE=1
+# Set environment variable to ensure output is sent straight to terminal without buffering
+ENV PYTHONUNBUFFERED=1
 
-CMD ["python", "bot.py"]
+# Sync the project into a new environment, asserting the lockfile is up to date
+RUN uv sync --locked
+
+CMD ["uv", "run", "whattsanddrops/main.py"]
