+++ pyproject.toml
@@ -10,6 +10,7 @@ requires-python = ">=3.8"
 readme = "README.md"
 dependencies = [
     "python-telegram-bot>=21.0",
+    "python-dotenv>=1.0.0",
 ]
 
 [project.optional-dependencies]
